requirements.txt
# Core
langchain==0.1.20
langchain-openai==0.0.8
langgraph==0.0.26
groq>=0.4.0
pydantic==2.6.1
pydantic-settings==2.1.0

# Utilidades
python-dotenv==1.0.1
rich==13.7.0
typer==0.9.0
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
